// src/app/core/models/customer.ts
// Generated by https://quicktype.io

export interface Customer {
  id?: number;
  firstName?: string;
  lastName?: string;
  email?: string;
  phoneNumber?: number;
  password?: string;
  gstNumber?: string;
  isactive?: boolean;
  createdOn?: string;
  createdBy?: number;
  updatedOn?: string;
  updatedBy?: number;
  addressId?: null;
  customerTypeId?: null;
  customerType?: null;
  cart?: any[];
  orders?: any[];
  wishlist?: any[];
}